requirements.txt
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.10
matplotlib==3.7.0
hydra-core==1.3.1
datasets==2.9.0
torchtext==0.14.1